backend/requirements.txt
# ============================================================
# Enterprise GuardCAP - Unified Backend Requirements
# ============================================================

# ===== FastAPI & Web Framework =====
fastapi>=0.109.0
uvicorn>=0.35.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ===== Database & Async =====
motor>=3.3.0
pymongo>=4.6.0
aiosmtpd>=1.4.4

# ===== Authentication & Security =====
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# ===== Environment & Config =====
python-dotenv>=1.1.0

# ===== NLP & RAG =====
torch>=2.0.0
transformers>=4.45.0
sentence-transformers>=3.0.0
chromadb>=0.5.3
konlpy>=0.6.0
numpy>=1.24.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2

# ===== LLM Integration =====
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# ===== PDF Processing =====
py-zerox>=0.0.7
openai>=1.12.0
pypdf>=4.0.0
pdf2image>=1.17.0
pillow>=10.0.0
pymupdf>=1.23.0

# ===== Data Processing & Utilities =====
requests>=2.31.0
tqdm>=4.66.0
jsonschema>=4.20.0
colorama>=0.4.6
rich>=13.7.0

# ===== Frontend (Optional) =====
# Flask>=2.0.0
# flask-cors>=3.0.10

# ===== UI (Optional - Streamlit) =====
# streamlit>=1.30.0

# ===== macOS Dependencies =====
certifi

# ===== Version Constraints =====
# urllib3 v2 경고 방지
urllib3<2.0.0